Move postUpgradeCommand execution to seperate file
