+++ lib/workers/branch/index.ts
@@ -1,6 +1,5 @@
 import is from '@sindresorhus/is';
 import { DateTime } from 'luxon';
-import minimatch from 'minimatch';
 import { RenovateConfig } from '../../config';
 import { getAdminConfig } from '../../config/admin';
 import {
@@ -15,36 +14,27 @@ import {
   TEMPORARY_ERROR,
   WORKER_FILE_UPDATE_FAILED,
 } from '../../constants/error-messages';
-import { addMeta, logger, removeMeta } from '../../logger';
+import { logger, removeMeta } from '../../logger';
 import { getAdditionalFiles } from '../../manager/npm/post-update';
 import { Pr, platform } from '../../platform';
 import { BranchStatus, PrState } from '../../types';
 import { ExternalHostError } from '../../types/errors/external-host-error';
 import { emojify } from '../../util/emoji';
-import { exec } from '../../util/exec';
-import { readLocalFile, writeLocalFile } from '../../util/fs';
 import {
   checkoutBranch,
   deleteBranch,
   getBranchCommit,
-  getRepoStatus,
   branchExists as gitBranchExists,
   isBranchModified,
 } from '../../util/git';
-import { regEx } from '../../util/regex';
-import * as template from '../../util/template';
 import { Limit, isLimitReached } from '../global/limits';
 import { checkAutoMerge, ensurePr, getPlatformPrOptions } from '../pr';
-import {
-  BranchConfig,
-  PrResult,
-  ProcessBranchResult,
-  BranchUpgradeConfig,
-} from '../types';
+import { BranchConfig, PrResult, ProcessBranchResult } from '../types';
 import { tryBranchAutomerge } from './automerge';
 import { prAlreadyExisted } from './check-existing';
 import { commitFilesToBranch } from './commit';
 import { getUpdatedPackageFiles } from './get-updated';
+import postUpgradeCommandExecutor from './post-upgrade-command-executer';
 import { shouldReuseExistingBranch } from './reuse';
 import { isScheduledNow } from './schedule';
 import { setStability } from './status-checks';
@@ -354,10 +344,7 @@ export async function processBranch(
       logger.debug('No updated lock files in branch');
     }
 
-    const {
-      allowedPostUpgradeCommands,
-      allowPostUpgradeCommandTemplating,
-    } = getAdminConfig();
+    const { allowedPostUpgradeCommands } = getAdminConfig();
 
     if (
       /* Only run post-upgrade tasks if there are changes to package files... */
@@ -367,226 +354,10 @@ export async function processBranch(
       getAdminConfig().trustLevel === 'high' &&
       is.nonEmptyArray(allowedPostUpgradeCommands)
     ) {
-      const executionMode = config.postUpgradeTasks.executionMode || 'update';
-
-      if (executionMode === 'branch') {
-        logger.trace(
-          {
-            tasks: config.postUpgradeTasks,
-            allowedCommands: allowedPostUpgradeCommands,
-          },
-          'Checking for post-upgrade tasks'
-        );
-        const commands = config.postUpgradeTasks.commands || [];
-        const fileFilters = config.postUpgradeTasks.fileFilters || [];
-
-        if (is.nonEmptyArray(commands)) {
-          // Persist updated files in file system so any executed commands can see them
-          for (const file of config.updatedPackageFiles.concat(
-            config.updatedArtifacts
-          )) {
-            if (file.name !== '|delete|') {
-              let contents;
-              if (typeof file.contents === 'string') {
-                contents = Buffer.from(file.contents);
-              } else {
-                contents = file.contents;
-              }
-              await writeLocalFile(file.name, contents);
-            }
-          }
-
-          for (const cmd of commands) {
-            if (
-              !allowedPostUpgradeCommands.some((pattern) =>
-                regEx(pattern).test(cmd)
-              )
-            ) {
-              logger.warn(
-                {
-                  cmd,
-                  allowedPostUpgradeCommands,
-                },
-                'Post-upgrade task did not match any on allowed list'
-              );
-            } else {
-              const execResult = await exec(cmd, {
-                cwd: config.localDir,
-              });
-
-              logger.debug(
-                { cmd, ...execResult },
-                'Executed branch level post-upgrade task'
-              );
-            }
-          }
-
-          const status = await getRepoStatus();
-
-          for (const relativePath of status.modified.concat(status.not_added)) {
-            for (const pattern of fileFilters) {
-              if (minimatch(relativePath, pattern)) {
-                logger.debug(
-                  { file: relativePath, pattern },
-                  'Post-upgrade file saved'
-                );
-                const existingContent = await readLocalFile(relativePath);
-                const existingUpdatedArtifacts = config.updatedArtifacts.find(
-                  (ua) => ua.name === relativePath
-                );
-                if (existingUpdatedArtifacts) {
-                  existingUpdatedArtifacts.contents = existingContent;
-                } else {
-                  config.updatedArtifacts.push({
-                    name: relativePath,
-                    contents: existingContent,
-                  });
-                }
-                // If the file is deleted by a previous post-update command, remove the deletion from updatedArtifacts
-                config.updatedArtifacts = config.updatedArtifacts.filter(
-                  (ua) => ua.name !== '|delete|' || ua.contents !== relativePath
-                );
-              }
-            }
-          }
-
-          for (const relativePath of status.deleted || []) {
-            for (const pattern of fileFilters) {
-              if (minimatch(relativePath, pattern)) {
-                logger.debug(
-                  { file: relativePath, pattern },
-                  'Post-upgrade file removed'
-                );
-                config.updatedArtifacts.push({
-                  name: '|delete|',
-                  contents: relativePath,
-                });
-                // If the file is created or modified by a previous post-update command, remove the modification from updatedArtifacts
-                config.updatedArtifacts = config.updatedArtifacts.filter(
-                  (ua) => ua.name !== relativePath
-                );
-              }
-            }
-          }
-        }
-      }
-
-      const filteredUpgradeCommands: BranchUpgradeConfig[] = config.upgrades.filter(
-        ({ postUpgradeTasks }) =>
-          !postUpgradeTasks ||
-          !postUpgradeTasks.executionMode ||
-          postUpgradeTasks.executionMode !== 'branch'
-      );
-
-      for (const upgrade of filteredUpgradeCommands) {
-        addMeta({ dep: upgrade.depName });
-        logger.trace(
-          {
-            tasks: upgrade.postUpgradeTasks,
-            allowedCommands: allowedPostUpgradeCommands,
-          },
-          'Checking for post-upgrade tasks'
-        );
-        const commands = upgrade.postUpgradeTasks.commands || [];
-        const fileFilters = upgrade.postUpgradeTasks.fileFilters || [];
-
-        if (is.nonEmptyArray(commands)) {
-          // Persist updated files in file system so any executed commands can see them
-          for (const file of config.updatedPackageFiles.concat(
-            config.updatedArtifacts
-          )) {
-            if (file.name !== '|delete|') {
-              let contents;
-              if (typeof file.contents === 'string') {
-                contents = Buffer.from(file.contents);
-              } else {
-                contents = file.contents;
-              }
-              await writeLocalFile(file.name, contents);
-            }
-          }
-
-          for (const cmd of commands) {
-            if (
-              allowedPostUpgradeCommands.some((pattern) =>
-                regEx(pattern).test(cmd)
-              )
-            ) {
-              const compiledCmd = allowPostUpgradeCommandTemplating
-                ? template.compile(cmd, upgrade)
-                : cmd;
-
-              logger.debug({ cmd: compiledCmd }, 'Executing post-upgrade task');
-
-              const execResult = await exec(compiledCmd, {
-                cwd: config.localDir,
-              });
-
-              logger.debug(
-                { cmd: compiledCmd, ...execResult },
-                'Executed post-upgrade task'
-              );
-            } else {
-              logger.warn(
-                {
-                  cmd,
-                  allowedPostUpgradeCommands,
-                },
-                'Post-upgrade task did not match any on allowed list'
-              );
-            }
-          }
-
-          const status = await getRepoStatus();
-
-          for (const relativePath of status.modified.concat(status.not_added)) {
-            for (const pattern of fileFilters) {
-              if (minimatch(relativePath, pattern)) {
-                logger.debug(
-                  { file: relativePath, pattern },
-                  'Post-upgrade file saved'
-                );
-                const existingContent = await readLocalFile(relativePath);
-                const existingUpdatedArtifacts = config.updatedArtifacts.find(
-                  (ua) => ua.name === relativePath
-                );
-                if (existingUpdatedArtifacts) {
-                  existingUpdatedArtifacts.contents = existingContent;
-                } else {
-                  config.updatedArtifacts.push({
-                    name: relativePath,
-                    contents: existingContent,
-                  });
-                }
-                // If the file is deleted by a previous post-update command, remove the deletion from updatedArtifacts
-                config.updatedArtifacts = config.updatedArtifacts.filter(
-                  (ua) => ua.name !== '|delete|' || ua.contents !== relativePath
-                );
-              }
-            }
-          }
-
-          for (const relativePath of status.deleted || []) {
-            for (const pattern of fileFilters) {
-              if (minimatch(relativePath, pattern)) {
-                logger.debug(
-                  { file: relativePath, pattern },
-                  'Post-upgrade file removed'
-                );
-                config.updatedArtifacts.push({
-                  name: '|delete|',
-                  contents: relativePath,
-                });
-                // If the file is created or modified by a previous post-update command, remove the modification from updatedArtifacts
-                config.updatedArtifacts = config.updatedArtifacts.filter(
-                  (ua) => ua.name !== relativePath
-                );
-              }
-            }
-          }
-        }
-      }
+      postUpgradeCommandExecutor('update', config);
+      postUpgradeCommandExecutor('branch', config);
     }
+
     removeMeta(['dep']);
 
     if (config.artifactErrors?.length) {
